+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog (nionui)
 0.3.28 (UNRELEASED)
 -------------------
 - Add backend support for section by section drawing for improved performance.
+- Add latency display (optional) to backend.
 - Fix drawing bugs (nested layers, startup race condition, thread shutdown).
 - Introduce action architecture for declarative menus and key bindings.
 - Add support for "name" keyword for declarative row and column widgets.